Krishna's Distribution
x^2cos(x) from 0 to pi/2

*I multiplied the pdf by 4/(pi^2-8) to normalize it
